Refactor Input component
