+++ src/Input/Input.constants.ts
@@ -1,0 +1,2 @@
+// Rendered height taking into account paddings, font-sizes and line-height
+export const height = 36
